added DELETE /roulette/:member_token/request/:request_id API.

+++ src/models/roulette-model.ts
@@ -6,6 +6,17 @@ import { ModelTypes } from './types';
 import { MysqlModules, MysqlTypes } from '../mysql';
 import { BaseLogicError } from '../errors';
 
+class RequestNotFoundError extends BaseLogicError {
+  constructor(msg: string) {
+    super('REQUEST_NOT_FOUND', msg);
+  }
+}
+class InvalidRouletteStatusError extends BaseLogicError {
+  constructor(msg: string) {
+    super('INVALID_ROULETTE_STATUS', msg);
+  }
+}
+
 injectable(ModelModules.Roulette.Cancel,
   [ MysqlModules.Mysql ],
   async (mysql: MysqlTypes.MysqlDriver): Promise<ModelTypes.Roulette.Cancel> =>
@@ -18,8 +29,10 @@ injectable(ModelModules.Roulette.Cancel,
           WHERE
             request_id=?
         `;
-        const resp = await mysql.query(requestSql, [ param.request_id ]);
-        console.log(resp);
+        const resp = await mysql.query(requestSql, [ param.request_id ]) as any;
+        if (resp.affectedRows !== 1) {
+          throw new RequestNotFoundError(`request not found with request_id: ${param.request_id}`);
+        }
 
         const checkerSql = `
           DELETE FROM
@@ -28,8 +41,10 @@ injectable(ModelModules.Roulette.Cancel,
             request_id=? AND
             match_status='WAITING'
         `;
-        const checkerResp = await mysql.query(checkerSql, [ param.request_id ]);
-        console.log(checkerResp);
+        const checkerResp = await mysql.query(checkerSql, [ param.request_id ]) as any;
+        if (checkerResp.affectedRows !== 1) {
+          throw new InvalidRouletteStatusError(`already matched request or request not found`);
+        }
       });
     });
